Move traverse logic to Program exit
I'm apologize for issuing this PR, because it's tailored to the use case
I have of running CommonJS modules in the browser. I've written a
plugin: https://github.com/tbranyen/babel-plugin-transform-commonjs
which converts CommonJS to browser-compatible ESM. I could only think of
two ways to use the two togther: 1) this change (or something similar to
it) is be made, 2) we wait until Babel implements a plugin ordering
system. I believe they are working on the latter.

My CommonJS plugin must run after all other transforms so that dead code
elimination can work properly. What this effectively does is halt the
traversal till the Program exit consistently. This will work reliably
for every plugin that does this, and ensures that all transforms have
run (and may run again) prior to running this plugin.

For now I'm going to publish a fork under my personal npm namespace so I
can keep moving, but I'd love to chat and figure out a way to either
land this or figure out a different solution.

Thanks in advance!